Move Storage and UI classes under recipeio package
